requirements.txt
bs4
python-telegram-bot
python-dotenv
img2pdf
aiohttp
pyTelegramBotAPI